backend/src/requirements.txt
absl-py==0.13.0
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
argon2-cffi==20.1.0
asgiref==3.4.1
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.2.0
Automat==20.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bleach==4.0.0
cachetools==4.2.2
certifi==2020.4.5.1
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.2
clang==5.0
click==7.1.2
cmake==3.21.1.post1
colorama==0.4.4
constantly==15.1.0
cors==1.0.1
crcmod==1.7
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.3
deepface==0.0.68
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.6
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-restframework==0.0.1
djangorestframework==3.12.4
dlib==19.22.1
dotmap==1.3.14
drf-extensions==0.7.1
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gdown==3.13.0
gevent==21.8.0
glob2==0.7
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.39.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
image==1.5.33
imageio==2.9.0
imutils==0.5.3
incremental==21.3.0
iniconfig==1.1.1
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaggle==1.5.12
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mtcnn==0.1.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.4.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
oss2==2.14.0
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.2.0
pluggy==0.13.1
priority==1.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
Protego==0.1.16
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyDispatcher==2.0.5
Pygments==2.10.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.3
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
queuelib==1.6.2
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rest-condition==1.0.3
retina-face==0.0.5
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scikit-plot==0.3.7
scipy==1.6.2
Scrapy==2.5.0
seaborn==0.11.2
Send2Trash==1.8.0
service-identity==21.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.11.0
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2021.3.31
tldextract==3.1.2
toml==0.10.2
tornado==6.1
tqdm==4.46.0
traitlets==5.0.5
Twisted==21.7.0
twisted-iocpsupport==1.0.1
typing-extensions==3.7.4.3
urllib3==1.25.11
virtualenv==20.6.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
